+++ examples/ipad_example.py
@@ -1,6 +1,6 @@
-from main import API
+from main import BaseAPI
 
-a=API()
+a=BaseAPI()
 a.sess='AAAATlrvMaj5B1tR4VrD1Qhy3GV_zukDoYg5KF_i6PTtcED04lhQItRHCj0A9Gx_RoAIHBvWLPZloWZfGGGt_6zzGqa1dHIO9cqZdIhdSc5cineU'
 a.uin='396184879'
 a.wait(5)#this will wait 5 seconds between each action
