--- items.py
@@ -1,4 +1,0 @@
-import json
-
-with open('data/items.json') as f:
-    data = json.load(f)
