--- stages.py
@@ -1,4 +1,0 @@
-import json
-
-with open('data/stages.json') as f:
-    data = json.load(f)
