--- weapon.py
@@ -1,4 +1,0 @@
-import json
-
-with open('data/weapons.json') as f:
-    data = json.load(f)
